improved the code, but not quite

+++ src/utils/builtin.c
@@ -46,7 +46,7 @@ void help(void)
     printf("-----------------------------------\n");
     printf("\nHere is the list of commands that you can execute:\n");
 
-    for (size_t i = 0; i < NbCommands; i++)
+    for (int i = 0; i < NbCommands; i++)
     {
         printf("%s\n", ComStr[i]);
     }
